Add workaround for a bug in ido
When both a default and an intial input are provided to ido, ido
places the point at the beginning of the input instead of at the end.
This advice works around the bug by preprocessing the arguments to
"ido-completing-read". If an initial input is provided and the default
is non-nil, we simply move the default item to the head of the choice
list, and then set the default to nil.

Note that this tweak is only applied when ido is used via the
ido-ubiquitous wrapper, so any direct calls to ido-completing-read
should be unaffected.

Fixes issue #18 in ido-ubiquitous.