Fix #19. Handle cons cell for initial-input arg
Instead of just nil or a string, the "initial-input" argument to
"completing-read" can also be a cons cell whose car is a string, and
my code didn't account for that, leading to spurious errors. Now the
code extracts the initial input string from the car of such a cons
cell.