Revert to from cl-loop to loop and require cl at runtime
cl-loop was introduced in 24.3, so for compatibility with earlier 24.x
versions we need to use just "loop".

The "eval-when-compile" trick seems too fragile, causing cryptic
errors that I dont understand, so I have removed it. These may be
related to some difference between the "cl-*" macros and their
unprefixed aliases. If someone submits a patch showing the correct way
to do the eval-when-compile trick while maintaining compatibility with
the entire version 24.x series, I'll happily merge it.