Fix #35
I think it is totally OK to defvar the variables for which the warnings are generated. Recompiling *.elc files after this change fixes the warnings.

From https://www.gnu.org/software/emacs/manual/html_node/eintr/defvar.html

"[defvar] only sets the value of the variable if the variable does not already have a value. If the variable already has a value, defvar does not override the existing value."

From defvar docstring:

"The optional argument INITVALUE is evaluated, and used to set SYMBOL,
only if SYMBOL's value is void.  If SYMBOL is buffer-local, its
default value is what is set; buffer-local values are not affected.
If INITVALUE is missing, SYMBOL's value is not set."