Remove old magit override on init
The old magit override was too broad. And now magit is switching to use
ido-ubiquitous, So users who have customized their overrides will not
have the old override auto-removed when they update, and magit could end
up stuck with no ido completion. So the init code now handles this
migration for the user and prompts them via warning to save the fix.