Depend on cl-lib, not cl-macs
In Emacs 24.1 and 24.2, cl-macs does not have a "provide" form, so it cannot be successfully `require`d. In any case, that file is an implementation detail, and client code should really just use `cl-lib`. This commit makes that change, and declares a package dependency on `cl-lib` so that we can rely on having sufficiently fixed `cl-*` functions/macros to hand.