silence byte-compiler
It appears that declaring a variable (using `(defvar variable)') does
not keep the byte-compiler from complaining when that variable is used
inside a `defadvice' form.  Use `with-no-warnings' instead.  As long as
`lexical-binding' is `nil' in all the involved libraries, that should be
safe.

Move the definition of `ido-cr+-debug-mode' near the beginning of the
buffer.  There was a comment that said that it was placed near the end
of the buffer so that it appears late in the Custom interface.  But I
believe that does not actually work, and even if it did that would not
have any effect because it is the *only* option, so being the first
option is the same as being the last option.