Default to only notifying about updates overrides
The new default for "ido-ubiquitous-auto-update-overrides" is "notify",
which will display a message using "display-warning" whenever updates
are available. This warning will happen on every startup until the user
customizes "ido-ubiquitous-auto-update-overrides" to t or nil, but now
the default will not modify the configuration.

this also more or less reverts the previous commit, since it is made
obsolete by this one.