Alternate handling of nil default
The new behavior is based on the principle that "empty string is the
default default" for completing-read, as explained here:
https://debbugs.gnu.org/cgi/bugreport.cgi?bug=27158#8

This should hopefully give somewhat reasonable behavior in all cases
without the need for caller-dependent behavior. However, there is
still a facility for overriding this behavior in order to improve the
situation for callers that don't want/expect completing-read to return
an empty string.