Make ido-cr+ handle dynamic completion tables
The idea for this implementation was copied from ivy-mode:
https://github.com/abo-abo/swiper

Note that this doesn't mean ido-cr+ can completely replace
completing-read-default everywhere. Things like tmm which implement
custom display and keybindings on top of completing-read will still
break it. Ido-ubiquitous may still be required as a gatekeeper for
these cases.

This now makes ido-describe-fns completely obsolete, so it has been
deleted.