Improvements to handling of dynamic completion tables
Now updating of the completion list for dynamic completion tables is
lazy whenever there are still multiple matches available. It is also
more thorough, getting completions for every prefix of ido-text as
well as the first match, rather than just checking ido-text and the
empty string.

Among other things, this should enable flex-matching to mostly work
with dynamic completion tables.

Tests are added for most of the new features.