Minor improvements to dynamic collection updating
Now we save time by avoiding dynamic updates when the new text is a
prefix of the text that was used for the previous dynamic update,
unless the user presses TAB to explicitly request more completions. We
also properly maintain the value of the "ido-rescan" variable through
dynamic updates, so that ido does not overzealously reorder the list
of completions without user input.

Fixes #153.